make CMapDEM::getElevation() and CMapDEM::getRegion() produce equal results
git-svn-id: http://svn.code.sf.net/p/qlandkartegt/code/QLandkarteGT/trunk@967 d0c010d8-bb82-477e-a075-112140230451

+++ src/CMapDEM.h
@@ -48,7 +48,7 @@ class CMapDEM : public IMap
         void zoom(qint32& level);
         void select(const QRect& rect);
         void dimensions(double& lon1, double& lat1, double& lon2, double& lat2);
-        float getElevation(float lon, float lat);
+        float getElevation(double lon, double lat);
         void getRegion(float *buffer, XY p1, XY p2, int w, int h);
 
     private:
